Made extensions into service
