+++ app/src/main/java/ch/abwesend/privatecontacts/application/KoinModule.kt
@@ -42,6 +42,7 @@ import ch.abwesend.privatecontacts.infrastructure.repository.ToastRepository
 import ch.abwesend.privatecontacts.infrastructure.repository.addressformatting.AddressFormattingRepository
 import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.repository.AndroidContactLoadRepository
 import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.repository.AndroidContactSaveRepository
+import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.service.AndroidContactChangesApplicationService
 import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.service.AndroidContactLoadService
 import ch.abwesend.privatecontacts.infrastructure.repository.androidcontacts.service.AndroidContactSaveService
 import ch.abwesend.privatecontacts.infrastructure.room.database.AppDatabase
@@ -77,6 +78,7 @@ internal val koinModule = module {
     single<IAndroidContactLoadService> { AndroidContactLoadService() }
     single { AndroidContactLoadService() }
     single<IAndroidContactSaveService> { AndroidContactSaveService() }
+    single { AndroidContactChangesApplicationService() }
 
     single { AndroidContactPermissionHelper() } // needs to be as singleton for initialization with the Activity
     single { CallPermissionHelper() } // needs to be as singleton for initialization with the Activity
